test: move text logo to components
