+++ src/app/components/TextLogo/Loadable.ts
@@ -1,0 +1,12 @@
+/**
+ *
+ * Asynchronously loads the component for TextLogo
+ *
+ */
+
+import { lazyLoad } from 'utils/loadable';
+
+export const TextLogo = lazyLoad(
+  () => import('./index'),
+  module => module.TextLogo,
+);
